requirements.txt
torch>=2.3.0
torchvision>=0.18.0
facenet-pytorch>=2.5.3
opencv-python>=4.10.0
numpy<2,>=1.26.0
pandas>=2.2.0
scipy>=1.11.0
Pillow>=10.3.0
tqdm>=4.66.0
typer[all]>=0.12.3
python-dotenv>=1.0.1
pyyaml>=6.0.1


fastapi>=0.111.0
uvicorn>=0.29.0
sqlmodel>=0.0.19
python-multipart>=0.0.9
